Repair history provider service
